python/requirements.txt
google-cloud-storage==2.14.0
google-cloud-firestore==2.14.0
google-cloud-aiplatform==1.38.1
google-cloud-documentai==2.20.1
pdfplumber==0.10.3
numpy==1.26.2